refactor: rename .new files to proper names
Remove inappropriate .new suffix from file names:
- MenuBar.new.tsx → MenuBar.tsx
- menu-new.css → menu.css

Updated all imports and references across:
- App.tsx: NewMenuBar → MenuBar
- components/menu/index.ts: Remove NewMenuBar alias
- css/index.css: menu-new.css → menu.css

Tests: 19/19 passed ✅
