+++ client/src/components/menu/index.ts
@@ -1,3 +1,2 @@
 export { MenuBar } from './MenuBar';
-export { MenuBar as NewMenuBar } from './MenuBar.new';
 export { StatusBar } from './StatusBar';
