+++ client/src/css/components/menu.css
@@ -1,93 +1,254 @@
+/* ============================================
+   MENU BAR - RStudio-Inspired Flat Design
+   ============================================
+   Design Philosophy:
+   - Flat, no gradients or blur
+   - Hard borders (1px solid)
+   - Compact spacing (32px height, 12px font)
+   - Monochrome base with biological accents
+   - Lab notebook yellow for AI features
+   ============================================ */
+
+/* --- Menu Bar Container --- */
 .menubar {
   display: flex;
   align-items: center;
   justify-content: space-between;
-  height: 44px;
-  padding: 0 20px;
-  background: linear-gradient(180deg, rgba(255, 255, 255, 0.9), rgba(244, 246, 251, 0.88));
-  border: 1px solid var(--border-color);
-  border-radius: var(--panel-radius);
-  backdrop-filter: blur(14px);
+  height: 32px; /* Reduced from 44px for RStudio compactness */
+  padding: 0 12px; /* Reduced from 20px */
+  background: #f6f8fa; /* Flat gray, no gradient */
+  border-bottom: 1px solid #d0d7de; /* Hard border */
+  border-radius: 0; /* No rounding */
+  box-shadow: none; /* No shadow */
+  backdrop-filter: none; /* No blur */
   position: relative;
-  overflow: hidden;
+  overflow: visible; /* Allow dropdowns */
 }
 
+/* Remove decorative overlay */
 .menubar::after {
-  content: '';
-  position: absolute;
-  inset: 0;
-  background: linear-gradient(120deg, rgba(47, 111, 237, 0.08), transparent 55%);
-  pointer-events: none;
+  display: none;
 }
 
 .menubar-left {
   display: flex;
   align-items: center;
   gap: 20px;
-  position: relative;
-  z-index: 1;
 }
 
 .menubar-brand {
-  font-size: 14px;
+  font-size: 13px; /* Slightly reduced from 14px */
   font-weight: 700;
   letter-spacing: -0.01em;
-  color: var(--accent-blue);
+  color: var(--text-primary);
+  font-family: var(--font-ui);
 }
 
 .menubar-menu {
   display: flex;
-  gap: 6px;
+  gap: 2px; /* Tighter spacing */
+}
+
+.menubar-right {
+  display: flex;
+  align-items: center;
+}
+
+/* --- Menu Items --- */
+.menu-section {
+  position: relative;
 }
 
 .menu-item {
-  padding: 6px 12px;
-  font-size: 12px;
+  padding: 4px 10px; /* Reduced from 6px 12px */
+  font-size: 12px; /* Compact */
   font-weight: 500;
   color: var(--text-secondary);
-  border-radius: 8px;
-  transition: all 0.15s ease;
+  border-radius: 3px; /* Subtle, not pill */
+  transition: none; /* No animations */
   background: transparent;
+  cursor: pointer;
+  position: relative;
+  border: none;
+  outline: none;
 }
 
 .menu-item:hover {
   color: var(--text-primary);
-  background: rgba(47, 111, 237, 0.12);
+  background: #e5e7eb; /* Flat gray, no blue tint */
 }
 
-.menubar-right {
-  position: relative;
-  z-index: 1;
+.menu-item:active,
+.menu-item.active {
+  background: #d1d5db; /* Slightly darker on click */
 }
 
+/* AI Assistant menu item - Lab notebook accent */
+.menu-item[data-id="edit:ai-assist"],
+.menu-dropdown-item[data-id="edit:ai-assist"] {
+  background: var(--lab-notebook, #fefce8); /* Lab notebook yellow */
+  border-left: 3px solid var(--dna-guanine, #f59e0b); /* DNA guanine orange */
+  padding-left: 7px; /* Adjust for border */
+  font-weight: 600;
+  color: var(--text-primary);
+}
+
+.menu-item[data-id="edit:ai-assist"]:hover,
+.menu-dropdown-item[data-id="edit:ai-assist"]:hover {
+  background: #fef9c3; /* Slightly darker yellow */
+}
+
+/* --- Connection Indicator --- */
 .connection-indicator {
   display: flex;
   align-items: center;
-  gap: 8px;
-  font-size: 12px;
+  gap: 6px; /* Reduced from 8px */
+  font-size: 11px; /* Reduced from 12px */
   color: var(--text-secondary);
-  padding: 6px 12px;
-  border-radius: 999px;
-  background: rgba(255, 255, 255, 0.6);
-  border: 1px solid rgba(148, 163, 184, 0.24);
-  box-shadow: inset 0 1px 0 rgba(255, 255, 255, 0.6);
+  padding: 4px 8px; /* Reduced from 6px 12px */
+  border-radius: 3px; /* Not pill (was 999px) */
+  background: #ffffff;
+  border: 1px solid #d0d7de;
+  box-shadow: none; /* No inset shadow */
 }
 
 .connection-dot {
-  width: 10px;
-  height: 10px;
+  width: 8px; /* Reduced from 10px */
+  height: 8px;
   border-radius: 50%;
-  background: var(--error-color);
-  box-shadow: 0 0 0 4px rgba(220, 76, 63, 0.12);
+  background: var(--error-color, #ef4444);
+  box-shadow: none; /* No glow effect */
 }
 
 .connection-indicator.connected .connection-dot {
-  background: var(--success-color);
-  box-shadow: 0 0 0 4px rgba(31, 146, 84, 0.2);
+  background: var(--success-color, #10b981);
+  box-shadow: none;
 }
 
 .connection-text {
-  font-size: 12px;
+  font-size: 11px;
   font-weight: 500;
   letter-spacing: 0.01em;
+}
+
+/* ============================================
+   DROPDOWN MENUS
+   ============================================ */
+
+.menu-dropdown {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  min-width: 220px;
+  background: #ffffff;
+  border: 1px solid #d0d7de;
+  border-radius: 0; /* No rounding */
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1); /* Subtle shadow only */
+  padding: 4px 0;
+  margin-top: 2px;
+  z-index: 1000;
+}
+
+.menu-dropdown-item {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 6px 12px;
+  font-size: 12px;
+  color: var(--text-primary);
+  cursor: pointer;
+  transition: none; /* No animation */
+  background: transparent;
+  border: none;
+  width: 100%;
+  text-align: left;
+  outline: none;
+}
+
+.menu-dropdown-item:hover {
+  background: #f6f8fa; /* Flat hover */
+}
+
+.menu-dropdown-item:active {
+  background: #e5e7eb;
+}
+
+.menu-dropdown-item[data-disabled="true"] {
+  color: var(--text-muted, #9ca3af);
+  cursor: not-allowed;
+}
+
+.menu-dropdown-item[data-disabled="true"]:hover {
+  background: transparent;
+}
+
+/* Separator */
+.menu-dropdown-separator {
+  height: 1px;
+  background: #e5e7eb;
+  margin: 4px 0;
+}
+
+/* Shortcut label */
+.menu-item-shortcut {
+  font-size: 11px;
+  color: var(--text-muted, #9ca3af);
+  font-family: var(--font-code, 'Berkeley Mono', 'Fira Code', 'Consolas', monospace);
+  margin-left: 24px;
+}
+
+/* Checkmark for toggle items */
+.menu-item-check {
+  width: 16px;
+  height: 16px;
+  margin-right: 8px;
+  color: var(--dna-adenine, #059669); /* Green checkmark */
+  visibility: hidden;
+}
+
+.menu-dropdown-item[data-checked="true"] .menu-item-check {
+  visibility: visible;
+}
+
+.menu-item-label {
+  display: flex;
+  align-items: center;
+}
+
+/* ============================================
+   CSS VARIABLES (if not already defined)
+   ============================================ */
+
+:root {
+  /* Base colors - RStudio-inspired */
+  --bg-primary: #ffffff;
+  --bg-secondary: #f6f8fa;
+  --bg-tertiary: #e5e7eb;
+  --border-color: #d0d7de;
+  --border-strong: #afb8c1;
+
+  /* Text */
+  --text-primary: #1f2937;
+  --text-secondary: #6b7280;
+  --text-muted: #9ca3af;
+
+  /* DNA bases (for accents) */
+  --dna-adenine: #059669;   /* Green - checkmarks */
+  --dna-thymine: #dc2626;   /* Red - errors */
+  --dna-cytosine: #2563eb;  /* Blue - info */
+  --dna-guanine: #f59e0b;   /* Yellow/Orange - warnings */
+
+  /* Lab environment */
+  --lab-notebook: #fefce8;  /* Yellow paper - AI features */
+  --lab-data: #f0f9ff;      /* Blue data sheets */
+
+  /* Functional colors */
+  --success-color: #10b981;
+  --error-color: #ef4444;
+  --warning-color: #f59e0b;
+  --info-color: #3b82f6;
+
+  /* Typography */
+  --font-ui: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
+  --font-code: 'Berkeley Mono', 'Fira Code', 'Consolas', monospace;
 }
