+++ client/src/css/index.css
@@ -8,6 +8,6 @@
 @import './components/plots.css';
 @import './components/console.css';
 @import './components/timeline.css';
-@import './components/menu-new.css';
+@import './components/menu.css';
 @import './components/status-bar.css';
 @import './components/layout.css';
